+++ gradle.properties
@@ -2,11 +2,10 @@
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 org.gradle.parallel=true
-org.gradle.workers.max=6
 org.gradle.caching=true
 #org.gradle.configuration-cache=true
 #org.gradle.configuration-cache.parallel=true
-version=1.3.0
+version=1.4.0
 group=com.github.tatercertified
 # Mod info
 mod_name=NoDim
@@ -18,22 +17,22 @@ homepage_url=https://github.com/Tater-Certified/NoDim#readme
 source_url=https://github.com/Tater-Certified/NoDim
 issue_url=https://github.com/Tater-Certified/NoDim/issues
 # Modloader info
-minecraft_version=1.21.10
-min_minecraft_version=1.14.3
-minecraft_version_range=[1.14.3,)
-parchment_minecraft=1.21.10
+minecraft_version=1.21.11
+min_minecraft_version=1.21.11
+minecraft_version_range=[1.21.11,)
+parchment_minecraft=1.21.11
 # https://parchmentmc.org/docs/getting-started
-parchment_version=2025.10.12
+parchment_version=2025.12.20
 # https://fabricmc.net/develop/
-fabric_loader_version=0.17.3
+fabric_loader_version=0.18.4
 # https://files.minecraftforge.net/net/minecraftforge/forge/
-forge_version=60.0.5
+forge_version=61.0.3
 forge_loader_version_range=[0,)
 # https://projects.neoforged.net/neoforged/neoforge
-neoforge_version=16-beta
+neoforge_version=14-beta
 neoforge_version_range=[0.0,)
 neoforge_loader_version_range=[0,)
 paper_version=R0.1-SNAPSHOT
 # https://spongepowered.org/downloads/spongevanilla
-sponge_version=17.0.0-SNAPSHOT
+sponge_version=18.0.0-SNAPSHOT
 java_version=21
