+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 spotless = "8.0.0"
 shadow = "9.2.2"
-unimined = "1.3.15"
+unimined = "8e64092954"
 
 [libraries]
 annotations = { group = "org.jetbrains", name = "annotations", version = "26.0.2" }
