+++ settings.gradle.kts
@@ -4,6 +4,17 @@ pluginManagement {
         gradlePluginPortal()
         mavenCentral()
         maven("https://maven.wagyourtail.xyz/releases")
+        maven { url = uri("https://jitpack.io") }
+    }
+
+    resolutionStrategy {
+        eachPlugin {
+            // Check if the requested plugin ID matches your target
+            if (requested.id.toString() == "xyz.wagyourtail.unimined") {
+                // Map it to the JitPack artifact coordinates
+                useModule("com.github.hypherionmc.Unimined:xyz.wagyourtail.unimined.gradle.plugin:8e64092954")
+            }
+        }
     }
 }
