+++ src/fabric/resources/nodim.mixins.fabric.json
@@ -8,22 +8,10 @@
   "plugin": "com.github.tatercertified.y_intmdry.NoDimMixinPlugin",
   "package": "com.github.tatercertified.y_intmdry.mixin",
   "mixins": [
-    "v1_14_3.BooleanValueMixin",
-    "v1_14_3.EndPortalMixin",
-    "v1_14_3.GameRuleRegistrationInterMixin",
-    "v1_14_3.GameRuleRegistrationSRGMixin",
-    "v1_14_3.GameRuleRegistrationMojMixin",
-    "v1_14_3.NetherPortalMixin",
-    "v1_14_3.TheEndGatewayMixin",
-    "v1_16.GameRuleRegistrationMixin",
-    "v1_17.TheGatewayPortalMixin",
-    "v1_21.GatewayPortalMixin",
-    "v1_21_5.EndPortalMixin",
-    "v1_21_5.GatewayPortalMixin",
-    "v1_21_5.NetherPortalMixin",
-    "v1_21_10.EndPortalMixin",
-    "v1_21_10.GatewayPortalMixin",
-    "v1_21_10.NetherPortalMixin"
+    "GameRuleRegistrationMixin",
+    "EndPortalMixin",
+    "GatewayPortalMixin",
+    "NetherPortalMixin"
   ],
   "client": [],
   "server": []
