+++ src/main/java/com/github/tatercertified/vanilla/NoDim.java
@@ -1,15 +1,15 @@
 /**
- * Copyright (c) 2025 QPCrummer
+ * Copyright (c) 2026 QPCrummer
  * This project is Licensed under <a href="https://github.com/Tater-Certified/NoDim/blob/main/LICENSE">MIT</a>
  */
 package com.github.tatercertified.vanilla;
 
-import net.minecraft.world.level.GameRules;
+import net.minecraft.world.level.gamerules.GameRule;
 
 public class NoDim {
-    public static final boolean DEBUG = false;
     public static final String MOD_ID = "nodim";
-    public static GameRules.Key<GameRules.BooleanValue> DISABLE_END;
-    public static GameRules.Key<GameRules.BooleanValue> DISABLE_NETHER;
-    public static GameRules.Key<GameRules.BooleanValue> DISABLE_GATEWAY;
+
+    public static GameRule<Boolean> DISABLE_END;
+    public static GameRule<Boolean> DISABLE_NETHER;
+    public static GameRule<Boolean> DISABLE_GATEWAY;
 }
