--- src/main/java/com/github/tatercertified/vanilla/annotation/MCVer.java
@@ -1,18 +1,0 @@
-/**
- * Copyright (c) 2025 QPCrummer
- * This project is Licensed under <a href="https://github.com/Tater-Certified/NoDim/blob/main/LICENSE">MIT</a>
- */
-package com.github.tatercertified.vanilla.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE})
-public @interface MCVer {
-    String min();
-
-    String max() default "";
-}
