+++ src/main/java/com/github/tatercertified/vanilla/mixin/GatewayPortalMixin.java
@@ -1,17 +1,16 @@
 /**
- * Copyright (c) 2025 QPCrummer
+ * Copyright (c) 2026 QPCrummer
  * This project is Licensed under <a href="https://github.com/Tater-Certified/NoDim/blob/main/LICENSE">MIT</a>
  */
-package com.github.tatercertified.vanilla.mixin.v1_21_10;
+package com.github.tatercertified.vanilla.mixin;
 
 import com.github.tatercertified.vanilla.NoDim;
-import com.github.tatercertified.vanilla.annotation.MCVer;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.EndPortalBlock;
+import net.minecraft.world.level.block.EndGatewayBlock;
 import net.minecraft.world.level.block.state.BlockState;
 
 import org.spongepowered.asm.mixin.Mixin;
@@ -20,19 +19,16 @@ import org.spongepowered.asm.mixin.injection.Coerce;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@MCVer(min = "1.21.10")
-@Mixin(EndPortalBlock.class)
-public class EndPortalMixin {
+@Mixin(EndGatewayBlock.class)
+public class GatewayPortalMixin {
     @Inject(
             method = {
-                "entityInside(Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/entity/Entity;Lnet/minecraft/world/entity/InsideBlockEffectApplier;Z)V", // Mojmap
-                "method_9548(Lnet/minecraft/class_2680;Lnet/minecraft/class_1937;Lnet/minecraft/class_2338;Lnet/minecraft/class_1297;Lnet/minecraft/class_10774;Z)V" // Intermediary
+                "entityInside(Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/entity/Entity;Lnet/minecraft/world/entity/InsideBlockEffectApplier;Z)V"
             },
             at = @At("HEAD"),
-            cancellable = true,
-            require = 0,
-            expect = 1)
-    private void nodim$checkIfEndIsEnabled(
+            cancellable = true
+    )
+    private void nodim$checkIfGatewayIsEnabled(
             BlockState blockState,
             Level level,
             BlockPos blockPos,
@@ -41,7 +37,7 @@ public class EndPortalMixin {
             boolean bl,
             CallbackInfo ci) {
         if (level instanceof ServerLevel serverWorld
-                && serverWorld.getServer().getGameRules().getBoolean(NoDim.DISABLE_END)) {
+                && serverWorld.getGameRules().get(NoDim.DISABLE_GATEWAY)) {
             ci.cancel();
         }
     }
