+++ src/main/java/com/github/tatercertified/vanilla/mixin/NetherPortalMixin.java
@@ -1,11 +1,10 @@
 /**
- * Copyright (c) 2025 QPCrummer
+ * Copyright (c) 2026 QPCrummer
  * This project is Licensed under <a href="https://github.com/Tater-Certified/NoDim/blob/main/LICENSE">MIT</a>
  */
-package com.github.tatercertified.vanilla.mixin.v1_14_3;
+package com.github.tatercertified.vanilla.mixin;
 
 import com.github.tatercertified.vanilla.NoDim;
-import com.github.tatercertified.vanilla.annotation.MCVer;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
@@ -16,24 +15,29 @@ import net.minecraft.world.level.block.state.BlockState;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Coerce;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@MCVer(min = "1.14.3", max = "1.21.4")
 @Mixin(NetherPortalBlock.class)
 public class NetherPortalMixin {
     @Inject(
             method = {
-                "entityInside", // Mojmap
-                "method_9548", // Intermediary
-                "m_7892_" // SRG
+                "entityInside(Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/entity/Entity;Lnet/minecraft/world/entity/InsideBlockEffectApplier;Z)V"
             },
             at = @At("HEAD"),
-            cancellable = true)
-    private void nodim$checkIfEndIsEnabled(
-            BlockState blockState, Level level, BlockPos blockPos, Entity entity, CallbackInfo ci) {
+            cancellable = true
+    )
+    private void nodim$checkIfNetherIsEnabled(
+            BlockState blockState,
+            Level level,
+            BlockPos blockPos,
+            Entity entity,
+            @Coerce Object insideBlockEffectApplier,
+            boolean bl,
+            CallbackInfo ci) {
         if (level instanceof ServerLevel serverWorld
-                && serverWorld.getServer().getGameRules().getBoolean(NoDim.DISABLE_NETHER)) {
+                && serverWorld.getGameRules().get(NoDim.DISABLE_NETHER)) {
             ci.cancel();
         }
     }
