--- src/main/java/com/github/tatercertified/vanilla/mixin/v1_14_3/BooleanValueMixin.java
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) 2025 QPCrummer
- * This project is Licensed under <a href="https://github.com/Tater-Certified/NoDim/blob/main/LICENSE">MIT</a>
- */
-package com.github.tatercertified.vanilla.mixin.v1_14_3;
-
-import com.github.tatercertified.vanilla.annotation.MCVer;
-
-import net.minecraft.world.level.GameRules;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Invoker;
-
-@MCVer(min = "1.14.3")
-@Mixin(GameRules.BooleanValue.class)
-public interface BooleanValueMixin {
-    @Invoker("create")
-    static GameRules.Type<GameRules.BooleanValue> create(boolean bl) {
-        throw new AssertionError();
-    }
-}
