--- src/main/java/com/github/tatercertified/vanilla/util/SerializedMCVer.java
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) 2025 QPCrummer
- * This project is Licensed under <a href="https://github.com/Tater-Certified/NoDim/blob/main/LICENSE">MIT</a>
- */
-package com.github.tatercertified.vanilla.util;
-
-public class SerializedMCVer {
-    private String id;
-
-    public String getId() {
-        return this.id;
-    }
-}
